+++ src/main/java/it/univaq/esc/dtoObjects/AppuntamentoDTO.java
@@ -16,7 +16,7 @@ import lombok.Setter;
 @Getter @Setter @NoArgsConstructor
 public class AppuntamentoDTO implements IModelToDTO{
     
-    private Integer idAppuntamento;
+    private Long idAppuntamento;
     private OrarioAppuntamento orarioAppuntamento = new OrarioAppuntamento();
     private List<UtentePolisportivaDTO> partecipanti = new ArrayList<UtentePolisportivaDTO>();
     private PrenotazioneSpecsDTO specificaPrenotazione;
