+++ src/main/java/it/univaq/esc/model/prenotazioni/RegistroAppuntamenti.java
@@ -184,7 +184,7 @@ public class RegistroAppuntamenti {
         return appuntamentiFiltrati;
     }
     
-    public Appuntamento getAppuntamentoById(Integer idAppuntamento){
+    public Appuntamento getAppuntamentoById(Long idAppuntamento){
         for(Appuntamento appuntamento: this.getListaAppuntamenti()){
             if(appuntamento.getIdAppuntamento() == idAppuntamento){
                 return appuntamento;
@@ -193,7 +193,7 @@ public class RegistroAppuntamenti {
         return null;
     }
     
-    public List<Appuntamento> getAppuntamentiByPrenotazioneId(Integer idPrenotazione){
+    public List<Appuntamento> getAppuntamentiByPrenotazioneId(Long idPrenotazione){
     	List<Appuntamento> appuntamenti = new ArrayList<Appuntamento>();
     	for(Appuntamento appuntamento : this.getListaAppuntamenti()) {
     		if(appuntamento.getIdPrenotazione() == idPrenotazione) {
